requirements.txt
streamlit
sentence-transformers==2.7.0
transformers==4.41.2
torch
faiss-cpu
numpy
ollama